+++ ember_mug/utils.py
@@ -45,12 +45,21 @@ def bytes_to_big_int(data: bytearray | bytes, signed: bool = False) -> int:
     return int.from_bytes(data, byteorder="big", signed=signed)
 
 
+def convert_temp_to_fahrenheit(temp: float) -> float:
+    """Convert to fahrenheit."""
+    return (temp * 9 / 5) + 32
+
+
+def convert_temp_to_celsius(temp: float) -> float:
+    """Convert to Celsius."""
+    return (temp - 32) * 5 / 9
+
+
 def temp_from_bytes(temp_bytes: bytearray, metric: bool = True) -> float:
     """Get temperature from bytearray and convert to Fahrenheit if needed."""
     temp = float(bytes_to_little_int(temp_bytes)) * 0.01
     if metric is False:
-        # Convert to fahrenheit
-        temp = (temp * 9 / 5) + 32
+        temp = convert_temp_to_fahrenheit(temp)
     return round(temp, 2)